requirements.txt
Flask==2.1.2
pyserial==3.5
